run "help" if no command given; remove no-pty
If the user does not run any command (e.g. `ssh git@host`) we now print
the help message.  If the user is not SSH'd, will print a message about
setting $SSH_ORIGINAL_COMMAND.

With this change, I take away the suggestion to put "no-pty" in the
authorized_keys.  If you leave this in and the user ssh's without
a command (as above), the output will be all messed up.  I don't think
there's any security risk with allowing PTYs.