fix for pod2usage(): cd to original cwd first
If you ran the script as a relative path (rather than an absolute path
or through $PATH), pod2usage() could not find the script since we
chdir'd at the start of the script.  The solution was to chdir after
potentially calling pod2usage(), and to chdir back to the original cwd
in &help().