added maximum path length check
Added a MAX_PATH_LEN (= 255) setting to Backend that makes sure that the
user-supplied path is of a maximum length.  I don't check that any
individual directory is within a certain length.  The number is
arbitrary, but I feel like 255 should be large enough for anyone.
I just don't want the users to make realy big paths - that just seems
wrong.