is_member(): deny if members file does not exist
Previously, if a user tried to access a repository in a group that did
not exist, an IOError was raised and not caught.  Now we deny if the
file is not found, since no such group exists.

Also removed the with statement, so we could get rid of that
requirement in the future.